Mona Alkhammash HW2
on a scale from 1 to 5, how comfortable were you with this assignment? Very Comfortable it was easy
What was a win you had with this assignment? Practies
What was a challenge you had with this assignment? thinking about an idea and styiling
Is there anything that you'd like some further information on? CSS styleing JS
Other stuff on your mind? work with anyone? tell us about it! \on a scale from 1 to 5, how comfortable were you with this assignment? Very Comfortable it was easy
What was a win you had with this assignment? Practies
What was a challenge you had with this assignment? thinking about an idea and styiling
Is there anything that you'd like some further information on? CSS styleing JS
Other stuff on your mind? work with anyone? tell us about it! :+1: